Fix installation of libnss_files
On glibc, dropbear seems to depend on the existence of libnss_files to
be able to read the password database and authorize logins. Rather than
hard-code the path and version, rely on the `inst_libdir_file` dracut
function, which should find the file if it exists and should fail
gracefully on other C libraries (like musl) where it doesn't. This is
how the build-in 95udev dracut module installs the library.

Also eliminate the reliance on `which` and improve error handling when
installing dropbear executable.

Closes #6.